fix a few bug: - vdpara string empty issue - check before image delete - delete PVD image file when delete vds record

+++ luhyacloud/clc/views.py
@@ -411,8 +411,11 @@ def buildUserObjectForScheduleRuleEngine(user):
     user["user_id"] = user_rec.userid
     user["user_name"]  = user_rec.showname
     user["user_group"] = user_rec.ec_authpath_name
-    vdpara = json.loads(user_rec.vdpara)
-    user["user_usb"]  = vdpara["usb"]
+    if len(user_rec.vdpara) > 0:
+        vdpara = json.loads(user_rec.vdpara)
+        user["user_usb"]  = vdpara["usb"]
+    else:
+        user["user_usb"] = "0"
     #user["user_superuser"] = auth_user_rec.is_superuser
 
     return user
@@ -999,9 +1002,13 @@ def edit_profile(request, uid):
         _vdpara = json.loads(ua.vdpara)
         _user['pds'] = _vdpara['pds']
         _user['sds'] = _vdpara['sds']
+        _user['vapp']= _vdpara['vapp']
+        _user['usb'] = _vdpara['usb']
     else:
-        _user['pds'] = ''
-        _user['sds'] = ''
+        _user['pds'] = "0"
+        _user['sds'] = "no"
+        _user['vapp']= "no"
+        _user['usb'] = "0"
 
     authnamelist =  ecAuthPath.objects.all()
     roles = []
@@ -1092,9 +1099,10 @@ def account_reset_password(request):
         user.save()
 
         ua = ecAccount.objects.get(userid=request.POST['userid'])
-        _vdpara = json.loads(ua.vdpara)
-        if _vdpara['vapp'] == 'yes':
-            virtapp_setPassword2AD(uid, newpw)
+        if len(ua.vdpara) > 0:
+            _vdpara = json.loads(ua.vdpara)
+            if _vdpara['vapp'] == 'yes':
+                virtapp_setPassword2AD(uid, newpw)
 
         response['Result'] = "OK"
         return HttpResponse(json.dumps(response), content_type='application/json')
@@ -2213,7 +2221,10 @@ def genRuntimeOptionForImageBuild(transid):
     user = tid_rec.user
     runtime_option["user"] = user
     user_obj = ecAccount.objects.get(userid=user)
-    user_vdpara = json.loads(user_obj.vdpara)
+    if len(user_obj.vdpara) > 0:
+        user_vdpara = json.loads(user_obj.vdpara)
+    else:
+        user_vdpara = {}
 
     ccobj       = ecServers.objects.get(ip0=ccip, role='cc')
     ccres_info  = ecCCResources.objects.get(ccmac0=ccobj.mac0)
@@ -2221,10 +2232,13 @@ def genRuntimeOptionForImageBuild(transid):
 
     # 0. get vm access protocol
     runtime_option['protocol'] =  getAccessProtocol(transid)
-    if user_vdpara['usb'] == '0':
-        runtime_option['usb_enabled'] = 0
+    if "usb" in user_vdpara.keys():
+        if user_vdpara['usb'] == '0':
+            runtime_option['usb_enabled'] = 0
+        else:
+            runtime_option['usb_enabled'] = 1
     else:
-        runtime_option['usb_enabled'] = 1
+        runtime_option['usb_enabled'] = 0
 
     # 1. general option
     img_info                        = ecImages.objects.get(ecid = src_imgid)
@@ -4511,6 +4525,13 @@ def delete_vds(request):
         response['Result'] = 'FAIL'
         response['errormsg'] = "Need to delete this VM's running task first"
     else:
+        # if PVD, need to delete image file
+        if vds_rec.insid.find("PVD") == 0:
+            _path = "/storage/pimage/%s/%s" % (vds_rec.user, vds_rec.imageid)
+            shutil.rmtree(_path)
+            if os.path.exists(_path):
+                logger.error("%s is not really deleted." % _path)
+
         # delete vds_auth records
         ecVDS_auth.objects.filter(insid=request.POST['insid']).delete()
         vds_rec.delete()
@@ -4802,20 +4823,25 @@ def list_images(request):
 def delete_images(request):
     response = {}
     rec = ecImages.objects.get(id=request.POST['id'])
+    tid_recs = ectaskTransaction.objects.filter(srcimgid=rec.ecid)
+    vds_recs = ecVDS.objects.filter(imageid=rec.ecid)
+    if tid_recs.count() != 0 or vds_recs.count() != 0:
+        response['Result'] = 'ERROR'
+        response['Message'] = "Need to delete this images's running TASKs and VDS first"
+    else:
+        ecImages_auth.objects.filter(ecid=rec.ecid).delete()
+        # delete image files
+        imgpath = '/storage/images/' + rec.ecid
+        if os.path.exists(imgpath):
+            shutil.rmtree(imgpath)
+        if rec.img_usage == 'server':
+            dbpath = '/storage/space/database/images/' + rec.ecid
+            if os.path.exists(dbpath):
+               shutil.rmtree(dbpath)
 
-    # delete image files
-    imgpath = '/storage/images/' + rec.ecid
-    if os.path.exists(imgpath):
-        shutil.rmtree(imgpath)
-    if rec.img_usage == 'server':
-        dbpath = '/storage/space/database/images/' + rec.ecid
-        if os.path.exists(dbpath):
-           shutil.rmtree(dbpath)
-
-    # delete image records
-    rec.delete()
-
-    response['Result'] = 'OK'
+        # delete image records
+        rec.delete()
+        response['Result'] = 'OK'
 
     retvalue = json.dumps(response)
     return HttpResponse(retvalue, content_type="application/json")
@@ -4995,8 +5021,9 @@ def update_active_account(request):
     u.save()
     ecu.save()
 
-    _vdpara = json.loads(ecu.vdpara)
-    virtapp_updateAccount2AD(u.username, _vdpara['vapp'])
+    if len(ecu.vdpara) > 0:
+        _vdpara = json.loads(ecu.vdpara)
+        virtapp_updateAccount2AD(u.username, _vdpara['vapp'])
 
     response['Result'] = 'OK'
     retvalue = json.dumps(response)
@@ -5753,9 +5780,10 @@ def verifySessionKey(session_key):
 def list_myvapps(uid):
     vapp = []
     userobj = ecAccount.objects.get(userid = uid)
-    para = json.loads(userobj.vdpara)
-    if para['vapp'] != 'yes':
-        return vapp
+    if len(userobj.vdpara) > 0:
+        para = json.loads(userobj.vdpara)
+        if para['vapp'] != 'yes':
+            return vapp
 
     # check vapp list available for this user
     return list_my_availed_vapp
@@ -5888,10 +5916,13 @@ def list_myvds(request):
     response['data'] = vds
     logger.error("user %s own virtual desktop as below: %s" %(_user, vds))
 
-    para = json.loads(ua.vdpara)
-    if para['vapp'] == 'yes':
-        myvapps = list_my_availed_vapp(_user)
-        response['vapp'] = myvapps['data']
+    if len(ua.vdpara) > 0:
+        para = json.loads(ua.vdpara)
+        if para['vapp'] == 'yes':
+            myvapps = list_my_availed_vapp(_user)
+            response['vapp'] = myvapps['data']
+        else:
+            response['vapp'] = []
     else:
         response['vapp'] = []
     logger.error("user %s own virtual app as below: %s" %(_user, response['vapp']))
